pyproject.toml
[project]
name = "apigw-http-api-lambda-dynamodb-python-cdk-stack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib>=2.245.0,<3.0.0",
    "boto3>=1.42.80",
    "constructs>=10.5.0,<11.0.0",
    "mypy-boto3-dynamodb>=1.42.73",

]

[dependency-groups]
dev = ["pytest==8.4.2", "types-boto3>=1.42.80"]
